Started on metadata, removed autoload, added libraries
A start was made on a configurable metadata scheme. For now, working with a PHP array,
to which fields and field parameters can be added. If enough fields are tested, and the
concept still holds, another structure outside the portal will be tought of (such as XML,
JSON or a custom designed language).
To allow for custom metadata schemes, the JavaScript was updated to be as general as
possible. If the edit for a field is cancelled, the value is returned to the original
value. If editing for the entire scheme is cancelled, all values are reset.

Autoload was removed from the portal in this commit, because it does not seem to work with
the HMVC plugin. This might be due to the fact that the MY_CONTROLLER core is used, rather
than the MX_Controller version provided by HMVC, but it might be something else as well.

Extra libraries were added, to allow functions to be shared between controllers. This applies
specifically to the modulelibrary library. From this library, module information can now
be requested (rather than from intake, which automatically loads a view)